+++ np.Templating/CHANGELOG.md
@@ -3,6 +3,13 @@
 ### About np.Templating Plugin
 See Plugin [README](https://github.com/NotePlan/plugins/blob/main/np.Templating/README.md) for details on available commands and use case.
 
+## [1.0.0-beta.33] - 2022-04-10 (mikeerickson)
+
+- fix issue passing renderData in `data` and `method` properties
+- Resolved issue when using `<%- discuss %>`
+  - Recommend refactoring to use `<%- prompt('discuss') %>`
+- Fixed issue with `<%- selection() %>` and `<%- system.selection() %>`
+
 ## [1.0.0-beta.32] - 2022-04-09 (mikeerickson)
 
 - Refactored template migration `np:migrate-templates` to create frontmatter templates